add genkey to dsaparam

+++ tests/dsa/dsa-test.sh
@@ -31,7 +31,55 @@ fi
 
 run_fail "dsaparam 0"
 run "dsaparam -out dsa.params 1024"
+run "dsaparam -in dsa.params"
+echo $RESULT | grep -e "-----BEGIN DSA PARAMETERS-----"
+if [ $? != 0 ]; then
+    echo "unexpected text output found"
+    echo "$RESULT"
+    exit 99
+fi
 
+#check for no output
+run "dsaparam -in dsa.params -noout"
+echo $RESULT | grep -e "-----BEGIN DSA PARAMETERS-----"
+if [ $? == 0 ]; then
+    echo "unexpected text output found"
+    echo "$RESULT"
+    exit 99
+fi
+
+#generate a key
+run "dsaparam -in dsa.params -genkey"
+echo $RESULT | grep -e "-----BEGIN DSA PARAMETERS-----"
+if [ $? != 0 ]; then
+    echo "unexpected text output found"
+    echo "$RESULT"
+    exit 99
+fi
+echo $RESULT | grep -e "-----BEGIN DSA PRIVATE KEY-----"
+if [ $? != 0 ]; then
+    echo "unexpected text output found"
+    echo "$RESULT"
+    exit 99
+fi
+
+#check noout with generate a key
+run "dsaparam -in dsa.params -genkey -noout"
+echo $RESULT | grep -e "-----BEGIN DSA PARAMETERS-----"
+if [ $? == 0 ]; then
+    echo "unexpected text output found"
+    echo "$RESULT"
+    exit 99
+fi
+echo $RESULT | grep -e "-----BEGIN DSA PRIVATE KEY-----"
+if [ $? != 0 ]; then
+    echo "unexpected text output found"
+    echo "$RESULT"
+    exit 99
+fi
+
+#check bad input
+run_fail "dsaparam -in ./certs/server-cert.pem -genkey -noout"
 rm -f dsa.params
 
 echo "Done"
